locust/requirements.txt
blinker==1.6.2
Brotli==1.0.9
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
ConfigArgParse==1.5.3
Flask==2.3.1
Flask-BasicAuth==0.2.0
Flask-Cors==3.0.10
gevent==22.10.2
geventhttpclient==2.0.9
greenlet==2.0.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
locust==2.15.1
MarkupSafe==2.1.2
msgpack==1.0.5
paho-mqtt==1.6.1
psutil==5.9.5
python-dotenv==1.0.0
pyzmq==25.0.2
requests==2.29.0
roundrobin==0.0.4
six==1.16.0
typing_extensions==4.5.0
urllib3==1.26.15
Werkzeug==2.3.2
zope.event==4.6
zope.interface==6.0
